// packages/theme/src/themes/Theme.type.ts
type ThemeType = {
  textColor: {
    DEFAULT: string;
  };
  lineHeight: {
    body: string;
    callout: string;
    footnote: string;
    caption1: string;
    caption2: string;
    headline: string;
    largeTitle: string;
    title1: string;
    title2: string;
    title3: string;
    subHeadline: string;
  };
  fontSize: {
    body: string;
    callout: string;
    footnote: string;
    caption1: string;
    caption2: string;
    headline: string;
    largeTitle: string;
    title1: string;
    title2: string;
    title3: string;
    subHeadline: string;
  };
  fontWeight: {};
  colors: {
    body: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    callout: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    footnote: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    caption1: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    caption2: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    headline: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    largeTitle: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    title1: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    title2: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    title3: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    subHeadline: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    label: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    fill: {
      primary: string;
      secondary: string;
      tertiary: string;
      quaternary: string;
    };
    background: {
      base: {
        primary: string;
        secondary: string;
        tertiary: string;
      };
      elevated: {
        primary: string;
        secondary: string;
        tertiary: string;
      };
    };
    materials: {
      background: {
        regular: string;
        thin: string;
        ultraThin: string;
        thick: string;
      };
    };
    groupedBackground: {
      primary: string;
      secondary: string;
      tertiary: string;
    };
    placeholderText: string;
    separator: string;
    opaqueSeparator: string;
    link: string;
    darkText: string;
    lightText: string;
    gray: {
      1: string;
      2: string;
      3: string;
      4: string;
      5: string;
      6: string;
    };
    red: {
      1: string;
    };
    orange: {
      1: string;
    };
    yellow: {
      1: string;
    };
    green: {
      1: string;
    };
    mint: {
      1: string;
    };
    teal: {
      1: string;
    };
    cyan: {
      1: string;
    };
    blue: {
      1: string;
    };
    indigo: {
      1: string;
    };
    purple: {
      1: string;
    };
    pink: {
      1: string;
    };
    brown: {
      1: string;
    };
  };
};

export type { ThemeType };